Address review: secure OAuth connect, consolidate provider labels
- Use fetch with Authorization header + redirect:manual instead of
  putting access_token in the URL query string (avoids credential leak
  via browser history, logs, and Referer headers). Falls back to query
  param for network errors or cross-origin opaque redirects.
- Extract PROVIDER_LABELS into shared lib/oauth-providers.ts utility
  and import from all three consuming components to avoid duplication.
